+++ mmyolo/models/dense_heads/yolov6_head.py
@@ -14,7 +14,7 @@ from mmengine.structures import InstanceData
 from torch import Tensor
 
 from mmyolo.registry import MODELS, TASK_UTILS
-from ...utils.misc import gt_instances_preprocess
+from ..utils import gt_instances_preprocess
 from .yolov5_head import YOLOv5Head
