+++ mmyolo/models/dense_heads/yolov8_head.py
@@ -14,8 +14,7 @@ from mmengine.structures import InstanceData
 from torch import Tensor
 
 from mmyolo.registry import MODELS, TASK_UTILS
-from ...utils.misc import gt_instances_preprocess
-from ..utils import make_divisible
+from ..utils import gt_instances_preprocess, make_divisible
 from .yolov5_head import YOLOv5Head
