+++ mmyolo/models/utils/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .misc import make_divisible, make_round
+from .misc import gt_instances_preprocess, make_divisible, make_round
 
-__all__ = ['make_divisible', 'make_round']
+__all__ = ['make_divisible', 'make_round', 'gt_instances_preprocess']
