+++ tests/test_models/test_utils/test_misc.py
@@ -6,8 +6,8 @@ import torch
 from mmengine.structures import InstanceData
 from torch import Tensor
 
+from mmyolo.models.utils import gt_instances_preprocess
 from mmyolo.utils import register_all_modules
-from mmyolo.utils.misc import gt_instances_preprocess
 
 register_all_modules()
